Metodos DAO's corregidos, filtración terminada y pene para todos

+++ src/main/java/com/coboard/dataaccess/dao/IProyectoDAO.java
@@ -1,5 +1,7 @@
 package com.coboard.dataaccess.dao;
 
+import java.math.BigInteger;
+
 import com.coboard.dataaccess.api.Dao;
 import com.coboard.modelo.Proyecto;
 
@@ -8,5 +10,5 @@ import com.coboard.modelo.Proyecto;
 *
 */
 public interface IProyectoDAO extends Dao<Proyecto, Integer> {
-	public Integer cantidadUsuariosPorProyecto (Proyecto proyecto);
+	public Long cantidadUsuariosPorProyecto (Proyecto proyecto);
 }
